Fix warning : implicit conversion changes signedness: 'long' to 'unsigned int' [-Wsign-conversion]

+++ modules/io/src/video/vpDiskGrabber.cpp
@@ -129,7 +129,7 @@ void vpDiskGrabber::acquire(vpImage<unsigned char> &I)
   m_image_number = m_image_number_next;
 
   if (m_use_generic_name) {
-    m_image_name = vpIoTools::formatString(m_generic_name, m_image_number);
+    m_image_name = vpIoTools::formatString(m_generic_name, static_cast<unsigned int>(m_image_number));
   }
   else {
     std::stringstream ss;
@@ -150,7 +150,7 @@ void vpDiskGrabber::acquire(vpImage<vpRGBa> &I)
   m_image_number = m_image_number_next;
 
   if (m_use_generic_name) {
-    m_image_name = vpIoTools::formatString(m_generic_name, m_image_number);
+    m_image_name = vpIoTools::formatString(m_generic_name, static_cast<unsigned int>(m_image_number));
   }
   else {
     std::stringstream ss;
@@ -172,7 +172,7 @@ void vpDiskGrabber::acquire(vpImage<float> &I)
   m_image_number = m_image_number_next;
 
   if (m_use_generic_name) {
-    m_image_name = vpIoTools::formatString(m_generic_name, m_image_number);
+    m_image_name = vpIoTools::formatString(m_generic_name, static_cast<unsigned int>(m_image_number));
   }
   else {
     std::stringstream ss;
@@ -194,7 +194,7 @@ void vpDiskGrabber::acquire(vpImage<unsigned char> &I, long image_number)
   m_image_number = image_number;
 
   if (m_use_generic_name) {
-    m_image_name = vpIoTools::formatString(m_generic_name, m_image_number);
+    m_image_name = vpIoTools::formatString(m_generic_name, static_cast<unsigned int>(m_image_number));
   }
   else {
     std::stringstream ss;
@@ -216,7 +216,7 @@ void vpDiskGrabber::acquire(vpImage<vpRGBa> &I, long image_number)
   m_image_number = image_number;
 
   if (m_use_generic_name) {
-    m_image_name = vpIoTools::formatString(m_generic_name, m_image_number);
+    m_image_name = vpIoTools::formatString(m_generic_name, static_cast<unsigned int>(m_image_number));
   }
   else {
     std::stringstream ss;
@@ -238,7 +238,7 @@ void vpDiskGrabber::acquire(vpImage<float> &I, long image_number)
   m_image_number = m_image_number_next;
 
   if (m_use_generic_name) {
-    m_image_name = vpIoTools::formatString(m_generic_name, m_image_number);
+    m_image_name = vpIoTools::formatString(m_generic_name, static_cast<unsigned int>(m_image_number));
   }
   else {
     std::stringstream ss;
